backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.2
PyJWT==2.12.1
bcrypt==4.0.1
python-dateutil==2.8.2
email-validator==2.1.0
pyotp==2.9.0
requests==2.31.0
cryptography==41.0.7
APScheduler==3.10.4
reportlab==4.0.8
pycryptodome==3.19.0
numpy==2.2.4
onnxruntime==1.21.0
Pillow==11.1.0
